Fix review comment
